changes to configs
