+++ chunkmydocs/src/utils/configs/mod.rs
@@ -1,9 +1,11 @@
 pub mod auth_config;
-pub mod extraction_config;
 pub mod github_config;
+pub mod llm_config;
 pub mod pdfium_config;
 pub mod rrq_config;
 pub mod s3_config;
+pub mod search_config;
 pub mod stripe_config;
 pub mod structured_extract;
 pub mod user_config;
+pub mod worker_config;
