+++ chunkmydocs/src/workers/postprocess.rs
@@ -1,10 +1,10 @@
-use chunkmydocs::utils::configs::extraction_config;
+use chunkmydocs::utils::configs::worker_config;
 use chunkmydocs::utils::rrq::consumer::consumer;
 use chunkmydocs::utils::workers::postprocess::process;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let config = extraction_config::Config::from_env()?;
+    let config = worker_config::Config::from_env()?;
     println!("Starting task processor");
     consumer(process, config.queue_postprocess, 1, 600).await?;
     Ok(())
